+++ delphi-frontend/src/test/resources/au/com/integradev/delphi/grammar/UnterminatedComment.pas
@@ -1,0 +1,9 @@
+unit UnterminatedComment;
+
+interface
+
+implementation
+
+{ oh no, this comment isn't terminated!
+
+end.